paid-blocks: refact -> split the code in two files
